Fix rebuildDrawingArea function
I misunderstood the requirements at first for this operation,
I thought it was to set each square's size to whatever the user
entered. But it was actually asking you to build the GRID size to
be whatever the user entered. So if the user said 8, it should build
an 8x8 grid of squares that fit into your defined container width.

This took some work and I learned the hard way that getting
[node].style.width does not actually give you the width of a node,
the only way to obtain that is either with [node].clientWidth or
[node].offsetWidth!